Provide FeldfreundConfiguration to System (#7)
To be more flexible with other projects, I added
`FeldfreundConfiguration` as a parameter to `System`. Before `System`
was loading the configuration itself during its `__init__`.

For that there a now the `config_from_id` and `config_from_file`
functions. `config_from_id` works exactly like before and
`config_from_file` takes a full path to the python configuration file.

I also added the `robot_id` to the config and renamed `name` to
`bluetooth_name`. And the pytest dependency was in wrong group.