Add ODrive `motor_temperature` to `TracksHardware` (#4)
We want to be able to see the ODrive temperature. For this, we use the
`motor_temperature` property from
[Lizard](https://github.com/zauberzeug/lizard/pull/174)￼. Lizard will
return `0` if there is no measurement. Since the chance that the actual
`motor_temperature` is `0` is very small, we will display `N/A` in this
case. This will help the user understand that the temperature sensor is
not working for their device.

---------

Co-authored-by: Pascal Schade <pascal@zauberzeug.com>